add js sources to ui:ui-graphics, commonize tests in ui:ui-graphics
