+++ compose/ui/ui-graphics/build.gradle
@@ -43,6 +43,7 @@ if(!AndroidXComposePlugin.isMultiplatformEnabled(project)) {
 
         testImplementation(libs.testRules)
         testImplementation(libs.testRunner)
+        testImplementation(libs.kotlinTestJunit)
         testImplementation(libs.junit)
         testImplementation(libs.truth)
 
@@ -68,6 +69,10 @@ if(AndroidXComposePlugin.isMultiplatformEnabled(project)) {
         iosX64("uikitX64")
         macosX64()
 
+        js(IR) {
+            browser()
+        }
+
         /*
          * When updating dependencies, make sure to make the an an analogous update in the
          * corresponding block above
@@ -90,6 +95,13 @@ if(AndroidXComposePlugin.isMultiplatformEnabled(project)) {
                 dependsOn(commonMain)
                 dependencies {
                     api(libs.skikoCommon)
+                }
+            }
+
+            jsMain {
+                dependsOn skiaMain
+                dependencies {
+                    implementation(libs.skikoJs)
                 }
             }
 
@@ -132,6 +144,15 @@ if(AndroidXComposePlugin.isMultiplatformEnabled(project)) {
                     implementation(libs.junit)
                     implementation(libs.truth)
                     implementation(libs.skikoCurrentOs)
+                }
+            }
+
+            commonTest {
+                kotlin.srcDir("src/test/kotlin")
+                dependencies {
+                    implementation(kotlin("test"))
+//                    implementation(libs.kotlinCoroutinesCore)
+//                    implementation("org.jetbrains.kotlinx:atomicfu:0.14.4")
                 }
             }
         }
