+++ compose/ui/ui-graphics/src/test/kotlin/androidx/compose/ui/graphics/ClipOpTest.kt
@@ -1,0 +1,29 @@
+/*
+ * Copyright (C) 2019 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.compose.ui.graphics
+
+import kotlin.test.assertEquals
+import kotlin.test.Test
+
+class ClipOpTest {
+
+    @Test
+    fun testToString() {
+        assertEquals("Difference", ClipOp.Difference.toString())
+        assertEquals("Intersect", ClipOp.Intersect.toString())
+    }
+}
