+++ compose/ui/ui-graphics/src/test/kotlin/androidx/compose/ui/graphics/StampedPathEffectStyleTest.kt
@@ -1,0 +1,30 @@
+/*
+ * Copyright (C) 2019 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.compose.ui.graphics
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class StampedPathEffectStyleTest {
+
+    @Test
+    fun testToString() {
+        assertEquals("Translate", StampedPathEffectStyle.Translate.toString())
+        assertEquals("Rotate", StampedPathEffectStyle.Rotate.toString())
+        assertEquals("Morph", StampedPathEffectStyle.Morph.toString())
+    }
+}
